sarkinmota_backend/requirements.txt
# ── Core ──────────────────────────────────────────────────────
Django==5.0.6
djangorestframework==3.15.2

# ── Authentication ────────────────────────────────────────────
djangorestframework-simplejwt==5.3.1

# ── CORS ──────────────────────────────────────────────────────
django-cors-headers==4.4.0

# ── Image uploads ─────────────────────────────────────────────
Pillow==10.3.0
cloudinary==1.40.0
django-cloudinary-storage==0.3.0

# ── Database & Deployment ─────────────────────────────────────
dj-database-url==2.2.0
psycopg2-binary==2.9.9
whitenoise==6.7.0
gunicorn==22.0.0

# ── Config management ─────────────────────────────────────────
python-decouple==3.8

# ── Filtering ─────────────────────────────────────────────────
django-filter==24.2

# ── API Docs ──────────────────────────────────────────────────
drf-spectacular==0.27.2

# ── Email (optional – SendGrid) ───────────────────────────────
sendgrid==6.11.0
